+++ Modules/Microsoft365DSC/Examples/Resources/AADAuthenticationMethodPolicyX509/2-Update.ps1
@@ -52,7 +52,7 @@ Configuration Example
                     TargetType = 'group'
                 }
             );
-            State                           = "disabled"; # Updated Property
+            State                           = "enabled";
         }
     }
 }